ml-service/requirements.txt
# Core ML Libraries
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scikit-learn==1.7.2
scipy>=1.11.0,<2.0.0

# Database
supabase==2.12.0
python-dotenv==1.0.0

# ML Models
xgboost>=2.0.0,<3.0.0
lightgbm>=4.1.0,<5.0.0
joblib>=1.3.0,<2.0.0

# Deep Learning (LSTM support for Premium tier)
torch>=2.1.0,<3.0.0

# Computer Vision (for pattern detection - optional, made optional in code)
opencv-python-headless>=4.8.0,<5.0.0

# API Server
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
pydantic>=2.5.0,<3.0.0

# Data Processing
python-dateutil>=2.8.0
pytz>=2023.3

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0

# Database - Supabase
supabase>=2.0.0
httpx>=0.24.0

# WebSocket support
websockets>=12.0

# Market data
yfinance>=0.2.31
requests>=2.31.0
aiohttp>=3.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
